requirements.txt
ckip-transformers==0.3.4
dateparser==1.2.1
faiss-cpu==1.10.0
flask==3.1.0
jieba==0.42.1
langchain==0.3.23
langchain-community==0.3.21
langchain-openai==0.3.12
openai==1.72.0
pandas==2.2.3
sentence-transformers==4.0.2
tiktoken==0.9.0
trafilatura==2.0.0
zhon==2.1.1